requirements.txt
cx-Oracle==8.0.1
google-cloud-bigquery==2.1.0